+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Added
   - Skipping texture UVs when there's no material using them
+- Fixed
+  - Computing position bounds
 
 ## [0.6.3] - 2019-10-17